Move DNSTAP parsing code to `lib/dnstap-parser`
