+++ lib/dnstap-parser/build.rs
@@ -1,0 +1,8 @@
+fn main() {
+    println!("cargo:rerun-if-changed=proto/dnstap.proto");
+    let mut prost_build = prost_build::Config::new();
+    prost_build.btree_map(["."]);
+    prost_build
+        .compile_protos(&["proto/dnstap.proto"], &[] as &[&str])
+        .unwrap();
+}
