+++ lib/dnstap-parser/src/vrl_functions/mod.rs
@@ -1,7 +1,7 @@
 use vrl::compiler::Function;
 
-mod parse_dnstap;
+pub mod parse_dnstap;
 
-pub(crate) fn all() -> Vec<Box<dyn Function>> {
+pub fn all() -> Vec<Box<dyn Function>> {
     vec![Box::new(parse_dnstap::ParseDnstap) as _]
 }
