+++ lib/vector-vrl/tests/Cargo.toml
@@ -7,6 +7,7 @@ publish = false
 
 [dependencies]
 chrono-tz.workspace = true
+dnstap-parser = { path = "../../dnstap-parser" }
 enrichment = { path = "../../enrichment" }
 vrl.workspace = true
 vector-vrl-functions = { path = "../../vector-vrl/functions" }
