+++ lib/vector-vrl/web-playground/Cargo.toml
@@ -13,6 +13,7 @@ wasm-bindgen = "0.2"
 vrl.workspace = true
 serde.workspace = true
 serde-wasm-bindgen = "0.6"
+dnstap-parser = { path = "../../dnstap-parser" }
 gloo-utils = { version = "0.2", features = ["serde"] }
 getrandom = { version = "0.2", features = ["js"] }
 vector-vrl-functions = { path = "../functions" }
