+++ src/sources/dnstap/mod.rs
@@ -2,38 +2,32 @@ use std::path::PathBuf;
 
 use base64::prelude::{Engine as _, BASE64_STANDARD};
 use dnsmsg_parser::dns_message_parser::DnsParserOptions;
+use dnstap_parser::parser::DnstapParser;
+use dnstap_parser::schema::DnstapEventSchema;
 use vector_lib::event::{Event, LogEvent};
 use vector_lib::internal_event::{
     ByteSize, BytesReceived, InternalEventHandle, Protocol, Registered,
 };
 use vector_lib::lookup::{owned_value_path, path};
 use vector_lib::{configurable::configurable_component, tls::MaybeTlsSettings};
-use vrl::compiler::Function;
 use vrl::path::{OwnedValuePath, PathPrefix};
 use vrl::value::{kind::Collection, Kind};
-
-use self::parser::DnstapParser;
 
 use super::util::framestream::{
     build_framestream_tcp_source, build_framestream_unix_source, FrameHandler,
 };
 use crate::internal_events::DnstapParseError;
-use crate::sources::dnstap::schema::DNSTAP_VALUE_PATHS;
 use crate::{
     config::{log_schema, DataType, SourceConfig, SourceContext, SourceOutput},
     Result,
 };
+use dnstap_parser::schema::DNSTAP_VALUE_PATHS;
 
-pub mod parser;
-pub mod schema;
 pub mod tcp;
 #[cfg(unix)]
 pub mod unix;
-use dnsmsg_parser::{dns_message, dns_message_parser};
-pub use schema::DnstapEventSchema;
 use vector_lib::config::{LegacyKey, LogNamespace};
 use vector_lib::lookup::lookup_v2::OptionalValuePath;
-mod vrl_functions;
 
 /// Configuration for the `dnstap` source.
 #[configurable_component(source("dnstap", "Collect DNS logs from a dnstap-compatible server."))]
@@ -348,10 +342,6 @@ impl FrameHandler for CommonFrameHandler {
     fn source_type_key(&self) -> Option<&vrl::path::OwnedValuePath> {
         self.source_type_key.as_ref()
     }
-}
-
-pub fn vrl_functions() -> Vec<Box<dyn Function>> {
-    vrl_functions::all()
 }
 
 #[cfg(test)]
