+++ src/sources/dnstap/unix.rs
@@ -11,7 +11,6 @@ use crate::{
     sources::util::framestream::UnixFrameHandler,
 };
 
-pub use super::schema::DnstapEventSchema;
 use vector_lib::EstimatedJsonEncodedSizeOf;
 
 /// Unix domain socket configuration for the `dnstap` source.
